create a src file and import function in page 2
I created an src folder with a data management file which contains functions to load  the clean and the feature engineered dataset in addition to a function which loads the pkl files. I then imported the function I needed in the page_2_correlation.py file and used it to access the feature engineered dataset for the dashboard.